+++ packages/cli/clients/directory/manageRegistration.go
@@ -24,7 +24,6 @@ import (
 	"github.com/cogment/cogment/grpcapi/cogment/api"
 	"github.com/cogment/cogment/utils"
 	"github.com/cogment/cogment/utils/endpoint"
-	"github.com/cogment/cogment/version"
 
 	"github.com/sirupsen/logrus"
 )
@@ -133,7 +132,6 @@ func ManageRegistration(
 	} else {
 		properties[endpoint.RegistrationSourcePropertyName] = "Self-Command_Line"
 	}
-	properties[endpoint.VersionPropertyName] = version.Version
 
 	log := log.WithFields(logrus.Fields{
 		"protocol": protocol,