+++ packages/cli/utils/endpoint/endpoint.go
@@ -34,7 +34,6 @@ const (
 	ServiceIDPropertyName           = "__id"
 	AuthenticationTokenPropertyName = "__authentication-token"
 	RegistrationSourcePropertyName  = "__registration_source"
-	VersionPropertyName             = "__version"
 )
 
 const servicePath = "service"