+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+-  Remove implicit directory property `__version` that contained invalid characters
+
 ## v2.19.0 - 2023-09-23
 
 ### Added